requirements.txt
Flask==2.0.1
numpy==1.21.5
Pillow==9.2.0
requests==2.27.1
Werkzeug==2.0.2
